Added package for mclapply to run on Windows.
Added the installation of the parallelsugar package and loaded the library at the start of the file. This checks what distribution system the user is operating in. If the user is using Mac or Linux, the default mclapply function in R is used. If the user is on Windows, a similar function is used within the parallelsugar package.